requirements.txt
streamlit==1.32.0
xgboost==2.0.3
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
rapidfuzz==3.6.1
firebase-admin==6.4.0
requests==2.31.0
python-dotenv==1.0.1
plotly==5.19.0